divided utility functions into multiple files
